Putting Babel deps into regular deps temporarily.
This is so that we can `install -g` from this github repo, without having to publish to npm yet. At the publishing step, we remove flow types using babel, but people want to try out the new `esy` stuff and we aren't yet ready to push a release to npm.