make esy use opam-alpha scope instead of opam
This branch can be checked out to make a new release which is configured
to use the opam-alpha scope name. Note that I had to update the
submodules to newly updated branches of camlp4/ocamlfind which also use
the opam-alpha scope. I will build a release and publish it as
esy#v0.0.2-alpha.

Hopefully very soon this can be merged into master.