[Esy] Fix missing opam-alpha conversion point.
Summary: We were using opam_ and we need to use opam_alpha_ for some of
the built in config. I choose to remove the constant config variable
because it gives the illusion that it was the single thing that needs to
be changed in order to change the prefix. This way, I'll make people
grep for it next time they want to change it (and they'll see they have
to make several changes across multiple languages).

Test Plan:

Reviewers:

CC: