Remove Root case for sourceType variant
The sourceType type is now convoluted, it is used for too many things at
once:

- How staleness is determined
- Which store to use for the build
- What's the root path for the package build

This commit attempts to simplify this a bit by removing the `Root` case
of `sourceType` (both from esy and esy-build-package packages). This way
the package tree can be reconfigured dynamically (replace the root and
so on).

The code branches for `Root` case were replaced with `isRoot` check
which compares `sandboxPath` with `sourcePath` of the build. Next
commits will review such checks and will try to refactor them away as
much as possible.