esy: add support for %{...}% syntax in command exprs
This commit adds support for opam's `%{...}%` variable syntax directly
as part of the command expression parser. This makes it possible to
preserve suhc syntax in package.json/esy.json manifests.

We will think if we want to keep it here but for the time being this is
how the translation would work with opam packages.