esy: automatically add DUNE_BUILD_DIR
Set it to `#{self.target_dir}`.

An alternative would be to add `DUNE_BUILD_DIR` only if there's `dune`
mentioned in `"esy.build"` config but that means if some package starts
to wrap `dune` invocation in some other command then it would break and
won't be obvious why. Also given that `DUNE_BUILD_DIR` contains `DUNE_`
prefix in its name it should be relatively safe to pollute build env
with this.