esyi: fix installation layout
Make it so `npm ls` doesn't report missing packages for `webpack` and
`react-scripts` installation.

We were missing recording positions for packages from insertion point
till the requested point.

Consider this setup:

- Package `pkg@1` is requested at `node_modules/a/node_modules/b/node_modules`
- Package `pkg@2` is requested at `node_modules/a/node_modules/c/node_modules`

Now if we insert `pkg@1` at `node_modules` we need to mark the following
locations as occupied between the actual insertion point and requested
point:

- `node_modules` (actual insertion point)`
- `node_modules/a/node_modules`
- `node_modules/a/node_modules/b/node_modules` (requested point)

This is because node module resolution algo which recursed up the tree.

Now the only way to install `pkg@2` is at

- `node_modules/a/node_modules/c/node_modules`