esyi: allow Resolver.package to fail with a reason
We make `Resolver.package` be:

    val package :
      resolution:Resolution.t
      -> t
    -  -> Package.t RunAsync.t
    +  -> (Package.t, string) result RunAsync.t

So that it can produce error in case package cannot be cretaed for a
given resolution (missing checksums and etc).

Solver now ignores such packages (and prints a warning) while naive
solver fails (should be ok for npm packages rather that surprising
backtrack to the prev. version).

Fixes #341