Fix for _esylink on Windows
We place _esylink before unpacking tarball, but that's just because we
get failures on Windows due to permission errors (reproducible on
AppVeyor).

I'd prefer to place _esylink after unpacking tarball to prevent tarball
contents overriding _esylink accidentially but probability of such event
is low enough so I proceeded with the current order.