Metadata overrides in resolutions (#451)
* Make Resolution a record

* Remove Package.source's SourceSpec case.

This was a hack previously which is not needed anymore.

* Drop `option` from `CommandList.t`

* Merge Resovler.Resolution and Package.Resolution, add override case

* Apply resolutions via resolver

* Fix NpmRegistryMock to mock package version requests

* Fix handling resolutions for npm packages

* Move SandboxSpec and EsyLinkFile to esyi

* Store override in _esylink file

* _esylink: optional override

* Update e2e tests

* Describe installation with _esylink files

Previously we used _esylink files to describe linked package, now we use
_esylink files for every installation.

This is needed as we want to communicate more info from installer to
builder, such as override. We'd also want to use that to communicate
`_esy/opam` metadata (but it's not done yet).

Dirs inside `node_modules` without `_esylink` file found will get
treatment as they are linked packages. Thus we preserve an ability to
"just copy package into node_modules dir".

* Change pkg deps repr

Storing a tagged list doesn't make much sense, also we don't keep info
if dep is optional or not, sandboxes deal with that already.

* Implement fetching of path: deps

* Fix parsing resolutions with overrides

* Handle dependency overrides in installer

* Make LocalPath/LocalPathLink refer by relative paths

* Allow packages without manifests if override is available

* Move SourceType/BuildType to esy-lib

* Add BuildType.AsInPackageJson

* Handle buildType overrides

* Solver: match sources with override

* Fix env commands (fix *-env ./pkg invocation)

* Fix installing bins of links

* Fix tests to account for relative paths to links

* Support buildEnv/exportedEnv overrides

* update esy.lock.json

* Add lockfile for .esytest

* Fix for _esylink on Windows

We place _esylink before unpacking tarball, but that's just because we
get failures on Windows due to permission errors (reproducible on
AppVeyor).

I'd prefer to place _esylink after unpacking tarball to prevent tarball
contents overriding _esylink accidentially but probability of such event
is low enough so I proceeded with the current order.