Remove toString from PRINTABLE signature
We already have show there and this caused a lot of boilerplate and
confusion on which one to use. I defaulted on show as this is what
`ppx_deriving` generates though I'd prefer to have toString.

I've ledt CudfName.toString as to be symmetrical to CudfName.ofString
and this is actually not for showing values but rather for conversion.