Resololution-aware matching for versions and sources
Previosuly we tried to keep enought info in specs so that they can be
matched against sources (rewrite git spec to include commit and etc).

Now we do matching only in context of the current resolve which provides
the world view (source spec -> source mapping, npm tag -> npm version
mapping and so on).