Convert source tests to use ppx_expect
We used ppx_inline_test exclusively before but they have their
downsides:

- We used `compare` to check for equality and this means that we needed
to expose type constructors to test code which isn't always
possible/desirable.

- Test can't be used with dune promote feature and thus each
modification to data repr resulte in a lot of manual modifications of
tests.

Now with ppx_expect with can run:

    % dune promote

and see test cases automatically updated.

For string repr we use `val sexp_of_t` and not `val show` as we want to
see the underlying data structure which is often incomprehensible with
`val show` repr.