Do not chmod symlinks when fixing permissions
We need to do `chmod` before unlinking files so that we can unlink
"read-only" files.

But we don't need to apply `chmod` on symlinks though as this will chmod
its target, not the link itself.

In case a link points to a directory this will remove `x` permission
from a directory and thus cause "Permission Errors" down the line.