Refactor "esy install" to be atomic (#538)
* Add RunAsync.List.mapAndWait/mapAndJoin

This commits adds two utilities to `RunAsync.List` module:

- `val RunAsync.List.mapAndJoin` - maps over a list with a promise
  returning function and wait for each promise to complete, optionally
  limit the concurrency.

- `RunAsync.List.mapAndWait` - same as previous but specialized for
  `()` returning functions, returns a `()` afterwards.

This helps with the common pattern where you need to run tasks
concurrently optionally with a limited concurrency:

    let queue = LwtTaskQueue.create ~concurrency:100 () in
    let f x = ... in
    RunAsync.List.waitAll
      (List.map ~f:(fun x -> LwtTaskQueue.submit queue (fun () -> f x)) xs)

which now can be written as:

    let f x = ... in
    RunAsync.List.mapAndWait ~f xs

* Move bin linking and lifecycle execution to FetchStorage

That way we can minimze the API of FetchStorage and better encapsulate
installation logic. We still require that packages are being installed
in a topo sort order (dependencies are installed before the package is
beign installed as package lifecycle might require dependencies).

* Make sure we have actual node in $PATH

Some env managers for node like to override `$PATH` (nodenv for
example). We workaround that by overriding `$PATH` back in pnp.js.

* Limit install concurrency

So that we don't fail on big dependency graphs.

* Remove source dir if lifecycle failed.

This makes it try once more on a second run.

* Do not fail if some subset of dependencies' bins are linked already.

* Atomic npm installs

* Fix linking devDependencies in pnp manifest

  * Add test cases which use node's require

* Resolve `*.cmd` executables on Windows

* Generate node.cmd bin on Windows

Related: #533