Refactor solution lock (#555)
Refactor how we "lock" solutions.

The idea of a "locked" solution is to reuse a set of found package versions w/o
running expensive metadata resolution and constraint solving. What's more
importantly - "lock" the known-to-be-working configuration of packages - this
allows to clone the dev repo of an esy project and be sure it's going to
compile.

Previously we used `esy.lock.json` to store "locked" solution.

The main downside is that we need to lock a lot more info that just package
versions - we need to store an entire opam metadata (which unfortunately changes
often and even can be taken offline). It doesn't make sense to store entire
files within a JSON structure - lockfiles should be readable and maintainable -
it should be possible to inspect a lock diff when reviewing changes to a
project.

Therefor we change lock to be a directoery `esy.lock` with the following
structure:

- `esy.lock/index.json` is an index of versions used in a solution.

- `esy.lock/opam/<name>.<version>` is a subset of opam metadata used in a
  solution.

- `esy.lock/opam-override/<name>.<version>` is a subset of opam overrides used
  in a solution.

In addition to that change this commit does the following:

- Remove support for source override

  We did use source tarball override before with an old installer. Now we
  correctly fallback to the opam index and so don't need that.

- Unifies opam overrides and regular esy overrides.

- Cleanups `OpamOverrides` module - we now assume that the only patterns for
  overrides are specific versions.

- Make sandbox cache keep track of non-existent (but expected) files.

  This will make sure we can invalidate cache in case installation didn't
  exist but then appeared.