Windows - Fix 'remote-https' is not a git command (#556)
This fixes several issues with `git` that were exposed by Azure CI.

We were using the globally installed version of `git`, as opposed to the version we bundle with Cygwin. This had the following issues:
- Later versions of windows git `>2.18.0` have errors with `https://`, which came across in our tests
- EOL issues (the default setting is to enable `autocrlf`, which checks out repos with `CRLF`/`/r/n` style line endings - causes problems with build scripts.

In addition, there were also some bugs with `esy-bash` - we were handling hardlinks properly in our bundle, but we were not rehydrating symlinks correctly. This led to cygwin-style symlinks being lost in the package step (which is we couldn't access the cert bundle). This is fixed by bringing in the latest esy-bash.