Fix lock checksum to be sensitive to dependencies/devDependencies (#606)
Previously if you move a dep from devDependencies to dependencies
lock checksum didn't detect this but it should because lock encodes
package graph.

The proper fix would be to disentange package graph from constraint
resolutions but this is not what we have today, instead we make checksum
sensitive to dependencies vs. devDependencies.