Fix out of bound access in fastreplacestring
We have seen segfaults on macOS/Azure and finally I was able to get a
good repro with @opam/octavius.

Firing up the offending command with lldb have shown the problem with
out of bound access (off by 1 error).

The fix is consistent with [1] and [2].

The original page which this algo refernces [3] has a typo apparently.

Thanks @IwanKaramazow for confirming the bug.

[1]: https://en.wikipedia.org/wiki/Rabin–Karp_algorithm
[2]: https://www.topcoder.com/community/competitive-programming/tutorials/introduction-to-string-searching-algorithms/
[3]: http://www-igm.univ-mlv.fr/~lecroq/string/node5.html