Ignore VCS dir when relocating sources and checking mtime (#621)
There's no point in doing that and including vcs dirs made those
processes slower. We are going to ignore them from now on.

There's possibility we can break some package which is being linked and
relocated but linked packages with relocation is so slow already, almost
unusable.  I'd suggest to change to an `"unsafe"` build-type for those
instead.