Fix package id to be sensitive to resolutions/opam metadata (#639)
This makes package id in installer sensitive to checksums of file we include in
esy.lock and to ressolutions specified in `"resolutions"` configuration.

Thus we fix a possible installation cache corrpution.

Build store corrpution was also possible due to build id not being sensitive to
patches from opam/override repos. This is handled now by keeping track of
package id in build id.