esy add: update lockfile with new checksum (#646)
Fixes #640 by updating the lockfile checksum after we update sandbox
manifest.

I don't really like this fix hence the name

```
val unsafeUpdatechecksum : ...
```

in `SolutionLock`.

but it's the least disturbing and quickest way to fix this issue and
make `esy add` usable.

We will consider changing lock format to check versions for constraints
instead of storing a checksum. But not now.