Fix race condition with multiple packages at the root (#654)
* Fix race condition with multiple packages at the root

Previsouly `esy-build-package` was setting up a symlink for a build dir
to sandbox root but that caused problems as `esy-build-package` had
wrong heueristics on what root package is (it tried to compare pkg root
path to sandbox root path which breaks when we have multiple packages
with the same root).

Now we move setting up a symlink to `esy` which knows better about which
package is the real root.

This is also in line with making `esy-build-package` dumber.

* Cleanup: do not expose SandboxPath.t in Task API

All usage sites convert them to `Path.t` immediately so it makes sense
to expose already converted paths. Makes for less boilerplate.