Correct handling for immutables builds with transient deps
Previously we handled such builds as transients but this is not correct
- _build-type builds are really not that correct to be performed on
cached sources.

Also we can skip staleness check for such builds.

We introduce new source type called `ImmutableWithTransientDependencies`
which is being built into local store but uses relocatio strategy not to
pollute the source dir which is cached.