Correctly handle devDependencies when planning build
We were trying to plan devDependencies along with the root build and
keep them away from scope by doing this check

    if PackageId.Set.mem id pkg.devDependencies
    then skip
    else ...

This is not so bad as `pkg.devDependencies` is a difference between
`"devDependencies"` and `"dependencies"` in `package.json`. This is why
it is so subtle - pakckages in `pkg.devDependencies` were not propagated
to the root because of this.

So if you haave packages which specify only `devDependencies.ocaml` then
even if its other dependencies depend on `ocaml` it won't see `ocamlc`
(and others in its build env):

    % esy b which ocamlc
    FAIL

Instead of this dumb check we now iterate over all devDependencies after
we build plan for the root.