Skip building linked packages for `esy CMD` invocation
When working with linked packages it is useful to be able to use run
commands  incommand environment w/o building linked packages: consider
the case where we have an error in a linked packages and still want to
run an editor to fix it.

This commit changes how `esy CMD` works: if only attempts to build
regular dependencies and shortcircuits as soon as it sees a linked
package.

The `esy build-dependencies` command also made to build only regular
dependencies but `--all` can be passed to build all dependencies.