esy-build.test: add build env snapshots
This commit adds snapshots for build env for various sandbox setups.

As build env is specific to an OS (because build IDs differ on different
OSes and also there are some diffs in env), unforunately these snapshots
are OS-dependent.

For now I commit only macOS-snapshots but going to add Linux-snapshots
and Windows-snapshots later.