Windows - Long Paths - Part 1: Prune PATH prior to build (#733)
Issue: We've had a few cases where we hit the PATH length limit (8192 characters) on Windows. This manifests as failure to resolve items in PATH (for example, we saw issues with ocamlopt failing to find x86_64-w64-mingw32-ar). In addition, we hit this sometimes in 'nested' esy execution contexts, like was addressed in #685

Introducing longer prefix paths exacerbates this issue, making it more likely for us to hit that PATH length limit.

Proposed Fix: Prune the PATH variable - there are a significant amount of empty directories, and we don't need to include those.