Add esy-install-npm-release
An implementation of bin/esyInstallRelease.js which is now uses node
"native" tar/gzip instead of shelling out to `tar` program.

This has upside as we now more portable (and can fix culprit with `tar`
on Windows) but the downside is that the dev is more complicated.

I've choosen not to integrate the build process for
`bin/esyInstallRelease.js` into the main dev workflow as we are changing
it very very rare. Instead the `esy-install-npm-release` holds the
yarn/browserify project which should be installed separately. Then the
`bin/esyInstallRelease.js` is a built artifact.

To update `bin/esyInstallRelease.js` one should run:

    % make bin/esyInstallRelease.js

We also use babel for `esy-install-npm-release` configured to work for
Node 4.x (tested manually on macOS).