Add 'esy build --install' invocation (#809)
* Add 'esy build --install' invocation

This instructs `esy build` command to install built artifacts. We don't
do that by default for perfomance reasons but in some cases it might be
handly - for example when you want to rebuild just a single package from
a dep graph and have it available immediately.

* Disable test for `esy build --install` on Windows

It requires bash for now.