Stronger check for installation.json validness
This makes check for stake installation.json stronger.

In some cases we use installation to loop over each packages and then
join it with metadata from solution. If there are some packages in
installation which do not exist in solution - we fail.

This makes sure we invalidate installation if it contains some packages
which are not in solution.

Also this dedupes the logic for installation staleness check by
providing a single function `Fetch.maybeInstallationOfSolution` which
returns `option(Installation.t)`.

Fixes #834