install: make DistStorage.fetchIntoCache transactional
Previously it was fetching into a dest dir directly.

This is problematic for the following reasons:

- It was possible to leave the dest dir in corrupted state (it wasn't
global cached corruption but removeing _esy was necessary)

- currently we don't dedup resolving metadata from sources (an another
problem on its own) it was possible to get a race condition

Now we fetch into a temp dir and then use `rename` (which is atomic) to
produce the dest dir.