fix signal handling on unix (#1191)
Using `Unix.create_process_env` starts a child process, but nothing is
setup to forward signals or interrupts. Making it impossible for the
started program to handle signal itself using `Sys.signal` or
`Sys.catch_break`. This is a minor inconvinience for CLI programs (e.g.
you can't implement a REPL that only exits when sending CTRL-C twice
like nodejs does), but this is a major problem for program that are
supposed to run in k8s.

K8s expects programs to keep serving requests when it sends a TERM
signal. The TERM signal is send before k8s makes sure that they won't
receive traffic anymore. It's expected that they mark themselves as not
ready, handle all open traffic, and shutdown when done. For this the
programs need to handle the TERM signal.