Deleting make-platform-release.js script(s)
It was a Node.js script copying the binaries and esyInstallRelease.js
The windows script used cygwin's tar and other commands to achieve the
same. (It also made a note of how .cygwin folder is skipped because it
would be too big)

Binaries are installed differently now. Slightly. These scripts are
not in use in the CI anywhere. Hence, removing. They'll be present in
VCS history anyways