Normalize paths in wrapper bin's environment
We started seeing the following error after it was necessary to export
esy-zlib in esy's exportedEnv.

```
        ../../libtool: line 7502: cd:
	C:npmprefixnode_modules@esy-nightlyesy3/i/esy_zlib-8f8db659/lib:
	No such file or directory
```

Clearly, the '\' path separator was getting escaped. This broke all
paths that got exported as a result of including esy-zlib in the
packages to be exported along with esy binary.

This commit normalises all paths with '/' character even on Windows