CI: use --no-env for esy npm-release on Linux
Because esy binary for Linux are statically linked inside the Docker
image and needs no sandboxing. On the other hand, when sandbox is
applied (even if unnecessary), the git and curl commands launched by
esy do not esy the system's global environment that contain dynamic
libraries needed.

We dont need this problem on MacOS. If we do, we should document that
zlib is now a global dependency for esy, just like git and curl. It's
because of zlib that we need a binary wrapper around esy.

Windows doesn't have libraries like zlib global installed on user
machines. git and curl dont seem to need any at the moment.