Improvements in how bubbled errors are reported
There were more than one way to print errors that are collected by RunAsync.contextf
and reported at the top where RunAsync.t is run through Lwt event loop.

It was also noticed that in Cli.re, error logs would be printed to the
screen. While it's okay for debug logs to be present all over the
codebase app, info and error logs could be centralised in one
place.